Update lib.read
When I was running the code in storing data lib.read,
The data came in buffer 
```
this was the error null and this was the data Buffer(13) [123, 34, 102, 111, 111, 34, 58, 34, …]
```
was the output to the previous code.  
  
I edited the code to in lib.read which i decoded the data buffer and sent it in callback.